+++ crates/sqlite-store/src/sync.rs
@@ -182,8 +182,14 @@ impl SqliteStore {
                     PublicAccountUpdate::Full(account) => {
                         Self::update_account_state(&db_tx, &mut smt_forest, account)?;
                     },
-                    PublicAccountUpdate::Patch { new_header, patch } => {
-                        Self::apply_sync_account_patch(&db_tx, &mut smt_forest, new_header, patch)?;
+                    PublicAccountUpdate::Patch { new_header, storage, vault } => {
+                        Self::apply_sync_account_patch(
+                            &db_tx,
+                            &mut smt_forest,
+                            new_header,
+                            storage,
+                            vault,
+                        )?;
                     },
                 }
             }